attachments/views.py
from rest_framework import viewsets

from attachments.models import Attachment
from attachments.serializers import AttachmentSerializer


class AttachmentViewSet(viewsets.ModelViewSet):
    queryset = Attachment.objects.all()
    serializer_class = AttachmentSerializer
